Lost Found, Request form, News and About student pages finished

+++ app/src/main/java/com/example/iutsummer/ui/user/IUTMainRepository.kt
@@ -1,6 +1,7 @@
 package com.example.iutsummer.ui.user
 
 import com.example.iutsummer.App
+import com.google.firebase.auth.EmailAuthProvider
 import com.google.firebase.auth.FirebaseAuth
 
 class IUTMainRepository(private val app:App) {
@@ -8,8 +9,23 @@ class IUTMainRepository(private val app:App) {
     private val auth = FirebaseAuth.getInstance()
     var userrr = auth.currentUser
 
+    /**
+     * Gets user from databse
+     */
+    val student = database.getUser()
+
+    /**
+     * Removes current user from firebase account
+     */
     fun dropStudent(){
-        database.drop()
-        userrr?.delete()
+        val email = student.value?.email
+        if(email!=null) {
+            val credential = EmailAuthProvider.getCredential(email, "1234567")
+            userrr!!.reauthenticate(credential).addOnCompleteListener {
+                userrr!!.delete().addOnCompleteListener {
+                    database.drop()
+                }
+            }
+        }
     }
 }
